fix(registry): type Batch terminal expiry arithmetic
Production immediate-cancel verification reached the freshly instrumented authority and exposed PostgreSQL's exact error: the file-input completion INSERT used two otherwise-untyped parameters in , so parsing failed with ambiguous unknown + unknown before the atomic hold release could commit. The same latent shape existed in inline completion.\n\nCast both timestamp and retention operands to bigint at the SQL boundary and pin that inference in the real PostgreSQL lifecycle matrix. No money, retention duration, output, or API semantics change.\n\nChecks: cargo test -p registry pg::gemini_batch_admission::tests::page_and_item_limits_are_exact; cargo check -p registry; rustfmt on changed files; git diff --check. Trusted-host validation runs the complete real PostgreSQL authority matrix.

+++ crates/registry/src/pg/gemini_batch_settlement.rs
@@ -102,7 +102,7 @@ pub(super) fn complete_job_if_terminal(
     if input_kind == "inline" {
         tx.execute(
             "UPDATE gemini_batch_jobs SET terminal_items_ts=COALESCE(terminal_items_ts,$2),\
-             completed_ts=COALESCE(completed_ts,$2),result_expiration_ts=COALESCE(result_expiration_ts,$2+$3),\
+             completed_ts=COALESCE(completed_ts,$2),result_expiration_ts=COALESCE(result_expiration_ts,$2::bigint+$3::bigint),\
              update_ts=GREATEST(update_ts,$2) WHERE job_id=$1",
             &[&job_id, &ts, &BATCH_RESULT_RETENTION_SECS],
         )?;
@@ -117,7 +117,7 @@ pub(super) fn complete_job_if_terminal(
         "INSERT INTO gemini_batch_files(file_id,account_id,display_name,mime_type,size_bytes,sha256_digest,\
          source_kind,state,storage_kind,create_ts,update_ts,expiration_ts,received_bytes,next_chunk_index,chunk_count)\
          VALUES($1,$2,'Gemini Batch output','application/jsonl',0,decode(repeat('00',32),'hex'),\
-         'batch_output','processing','chunked',$3,$3,$3+$4,0,0,0) ON CONFLICT(file_id) DO NOTHING",
+         'batch_output','processing','chunked',$3,$3,$3::bigint+$4::bigint,0,0,0) ON CONFLICT(file_id) DO NOTHING",
         &[&file_id, &account_id, &ts, &BATCH_RESULT_RETENTION_SECS],
     )?;
     tx.execute(
